requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.1
psycopg2-binary==2.9.3
django-storages==1.12.3
boto3==1.24.0